requirements.txt
requests>=2.28
python-dotenv>=1.0
jinja2>=3.0
openai>=1.0.0
flask>=2.0
flask-cors>=3.0
pytest>=7.0
